More conventional naming and properly wrapping.
In normal Python code an argument taking decorator has two inner
functions:

- one that returns a parameterized decorator
- one that actually wraps the function being decorated.

As it is generally good to express your intention in code, the names of
these functions should be named accordingly. Hence the first one has
been changed to `decorate` and the second one to `wrapper`.

In addition it is good practise to invoke `functools.update_wrapper()`
when defining a wrapper function. This is easily done using the
`functools.wraps()` decorator. This preserves the wrapped function's
`__module__`, `__name__`, `__qualname__`, `__annotations__` and
`__doc__` attributes. For one this helps with debugging when using these
extremely unpythonic, heavily Scala inspired, constructs.