Fixed all remaining type annotation issues.
MyPy now runs cleany on the code base.

The @do decorator needed to be made more generic (less strict type
checking). This is due to the nature of how generators work. Generators
can `return` in addition to `yield`. When they return they raise a
`StopIteration`. The value returned will be in the attribute `value` of
the StopIteration exception. This value can be anything for the @do
decorator. Hence most type constraints using `TypeVar('T', Maybe,
Either)` had to be removed.

With TypeVar `T` constraint to `Maybe` and `Either` they needed to be
made callable (due to `M.unit(e.value)` call). Hence these abstract base
classes gained a proper `__init__` method.

Also fixed a number of issues in the `dict` and `list` module.