Don't use regex on structured file format!!!
Seriously we're dealing with YAML files; that is structured data
that's easily read in to be processed as a Python dictionaries. No
need for regexes to change the values of top level keys. Use Pyhon
dictionary syntax!

Reason for this change is that the regex contained an incorrect
assumption. It assumed that strings were double quoted. Thing is in
YAML string don't need to be quoted at all. Stupid assumption. Stupid
waste of my time.

Could have fixed the regex, but figured processing structured data in a
structured way is the better solution.