Move to Python 3.7 and remove quotes from type annotations.
This is not necessary anymore due to the introduction in Python 3.7 of
PEP 563, postponed evaluation of type annotations.

However this needed an additional config change/addition as the
`flake8-mypy` plugin internally specifies that it should run according
to Python 3.6 specs. By explicitely pointing it to an `mypy` config this
can be overriden. This config needs to be specific for `flake8-mypy`,
and hence can't be shared with `mypy`, as `flake8` checks things per
file, whereas `mypy` can checks things across file bounderies.  Note:
this has already been solved on the `master` branch of `flake8-mypy` but
is not yet part of a release. Once a new release is out we can ditch
`flake8_mypy.ini` and the reference to it from `setup.cfg'